requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.17.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
